Adding Stock quantity and related logic in progres

+++ src/main/java/com/example/bookeshop/models/Basket.java
@@ -19,14 +19,14 @@ public class Basket {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private int totalPrice;
+    private Long totalPrice;
 
     @OneToMany(cascade = CascadeType.ALL, fetch = FetchType.LAZY)
     private List<BasketItem> items = new ArrayList<>();
 
     public int calculateTotalPrice() {
         return items.stream()
-                .mapToInt(item -> item.getBook().getPrice() * item.getQuantity())
+                .mapToInt(item -> Math.toIntExact(item.getBook().getPrice() * item.getQuantity()))
                 .sum();
     }
 }
